tq-98: Refactor to EventAnalytic
