--- src/pages/EventStatistic/statisticElements/eventStatisticElementProps.ts
@@ -1,7 +1,0 @@
-import { UserProfile } from '../../../models/userProfile'
-import { FeedbackAnswerGetDto } from '../../../services/api/commonModule/commonEntities/feedback/feedbackDto'
-
-export interface EventStatisticElementProps {
-  participants: UserProfile[]
-  feedbackAnswers: FeedbackAnswerGetDto[]
-}
